backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
python-dotenv==1.1.0
google-generativeai>=0.7.2
sqlalchemy==2.0.41
aiosqlite==0.21.0
faker==37.3.0
pillow>=11.0.0
httpx==0.28.1
websockets==15.0.1
python-multipart==0.0.20
